requirements.txt
Flask
python-dotenv
pymongo[srv]
certifi
google-cloud-vision